AiImageProcessing/requirements.txt
SQLAlchemy~=2.0.38
passlib~=1.7.4
fastapi~=0.115.11
pydantic~=2.10.6
requests~=2.32.3
pip~=24.2
wheel~=0.43.0
pillow~=11.0.0
filelock~=3.17.0
torch~=2.6.0
torchvision~=0.21.0
matplotlib~=3.9.2
pandas~=2.2.3
scikit-learn~=1.6.1
pyodbc~=5.2.0